% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-methods.R
\name{approxInference}
\alias{approxInference}
\title{Importance Sampling (IS) in Bayesian Networks}
\usage{
approxInference(
  BayesNet,
  obs,
  s_method = "SGS",
  N_samples = 100,
  relevantSubNet = FALSE,
  plot = TRUE,
  returnList = FALSE
)
}
\arguments{
\item{BayesNet}{Bayesian network}

\item{obs}{List containing the evidence nodes and associated values}

\item{s_method}{Inference method}

\item{N_samples}{Number of samples}

\item{relevantSubNet}{Whether to consider the relevant subnet}

\item{plot}{If TRUE, plot results}

\item{returnList}{If TRUE, return detailed list of intermediate results}
}
\value{
Normalizing constant and vector of intermediate results
}
\description{
Outputs the normalizing constant given some evidence of a Bayesian network
}
