% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_mv}
\alias{arma_mv}
\title{Matrix Product in RcppArmadillo.}
\usage{
arma_mv(m, v)
}
\arguments{
\item{m}{numeric matrix}

\item{v}{numeric vector}
}
\value{
matrix product of m and v
}
\description{
Matrix Product in RcppArmadillo.
}
