% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{arma_sparse}
\alias{arma_sparse}
\title{Matrix Product in RcppArmadillo.}
\usage{
arma_sparse(m, v)
}
\arguments{
\item{m}{numeric sparse matrix}

\item{v}{numeric vector}
}
\value{
matrix product of m and v
}
\description{
Matrix Product in RcppArmadillo.
}
