% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asiaadj.R
\docType{data}
\name{Asiamat}
\alias{Asiamat}
\title{Asiamat}
\format{
A binary matrix with 8 rows and 8 columns representing an adjacency matrix of a DAG with 8 nodes:
\itemize{
\item D (dyspnoea), binary 1/0 corresponding to "yes" and "no"
\item T (tuberculosis), binary 1/0 corresponding to "yes" and "no"
\item L (lung cancer), binary 1/0 corresponding to "yes" and "no"
\item B (bronchitis), binary 1/0 corresponding to "yes" and "no"
\item A (visit to Asia), binary 1/0 corresponding to "yes" and "no"
\item S (smoking), binary 1/0 corresponding to "yes" and "no"
\item X (chest X-ray), binary 1/0 corresponding to "yes" and "no"
\item E (tuberculosis versus lung cancer/bronchitis), binary 1/0 corresponding to "yes" and "no"
}
}
\source{
\url{https://www.bnlearn.com/bnrepository/}
}
\usage{
Asiamat
}
\description{
An adjacency matrix representing the ground truth DAG used to generate a synthetic dataset from 
Lauritzen and Spiegelhalter (1988) about lung
diseases (tuberculosis, lung cancer or bronchitis) and visits to Asia.
}
\references{
Lauritzen S, Spiegelhalter D (1988). `Local Computation with Probabilities on Graphical Structures and their Application to Expert Systems (with discussion)'.
Journal of the Royal Statistical Society: Series B 50, 157-224.
}
\keyword{datasets}
