% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-methods.R
\name{benchmarkMultipleNets}
\alias{benchmarkMultipleNets}
\title{Benchmark Inference Methods}
\usage{
benchmarkMultipleNets(
  N_var,
  N_Obs,
  N_nets,
  N_rep,
  N_samples = 100,
  samplingMethods = c("GS", "SGS", "LBP"),
  DAGmethod = "er",
  N_neighbours = 2,
  nodeDim = 2
)
}
\arguments{
\item{N_var}{number of variables}

\item{N_Obs}{number of observations}

\item{N_nets}{number of Bayes nets}

\item{N_rep}{number of repetitions}

\item{N_samples}{number of samples}

\item{samplingMethods}{benchmarked sampling methods}

\item{DAGmethod}{DAG method}

\item{N_neighbours}{number of neighbours}

\item{nodeDim}{category size}
}
\value{
benchmark results
}
\description{
Outputs benchmark results of different inference methods
}
