% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion_BiDAG_Bestie-methods.R
\name{convertBNBestieToSGS}
\alias{convertBNBestieToSGS}
\title{Bayesian network Conversion Bestie to SGS format}
\usage{
convertBNBestieToSGS(DAGaugmented)
}
\arguments{
\item{DAGaugmented}{Bayes net of type Bestie (created by DAGparameters())}
}
\value{
Bayes net of SGS format
}
\description{
Converts Bayesian network of Bestie format to SGS format
CPTs of Bestie are created by DAGparameters() 
and DAG of Bestie is created by BiDAG::iterativeMCMC()$DAG
}
