% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion_BiDAG_Bestie-methods.R
\name{convertCPTsBestieToSGS}
\alias{convertCPTsBestieToSGS}
\title{CPTs Conversion Bestie to SGS format}
\usage{
convertCPTsBestieToSGS(DAGaugmented)
}
\arguments{
\item{DAGaugmented}{CPTs of type Bestie (created by Bestie:::DAGparameters())}
}
\value{
CPTs of SGS format
}
\description{
Converts CPTs of Bestie format to SGS format
}
