% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion_BiDAG_Bestie-methods.R
\name{DAGparameters}
\alias{DAGparameters}
\title{Learn CPTs (taken from Bestie package)}
\usage{
DAGparameters(incidence, dataParams)
}
\arguments{
\item{incidence}{DAG}

\item{dataParams}{score parameters}
}
\value{
CPTs of the Bayes net
}
\description{
Return the CPTs for a Bayesian network (author Jack Kuipers)
}
