% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgsR.R
\name{findMatrix}
\alias{findMatrix}
\title{Findmatrix type}
\usage{
findMatrix(object, matrix.slot)
}
\arguments{
\item{object}{Seurat object}

\item{matrix.slot}{the name of the slot}
}
\value{
Return a matrix object from a Seurat object or show an error message
}
\description{
find expression matrix from seurat object like:counts,scale.data
}
