% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sgs.R
\name{fit_sgs}
\alias{fit_sgs}
\title{fit an SGS model}
\usage{
fit_sgs(
  X,
  y,
  groups,
  pen_method = 1,
  type = "linear",
  lambda,
  alpha = 0.95,
  vFDR = 0.1,
  gFDR = 0.1,
  max_iter = 5000,
  backtracking = 0.7,
  max_iter_backtracking = 100,
  tol = 1e-05,
  standardise = "l2",
  intercept = TRUE,
  w_weights = NULL,
  v_weights = NULL,
  x0 = NULL,
  u = NULL,
  verbose = FALSE
)
}
\arguments{
\item{X}{Input matrix of dimensions \eqn{p \times n}.}

\item{y}{Output vector of dimension \eqn{n}.}

\item{groups}{A grouping structure for the input data. Should take the form of a vector of group indices.}

\item{pen_method}{The type of penalty sequences to use.
\itemize{
\item \code{1} uses the vMean SGS and gMean gSLOPE sequences.
\item \code{2} uses the vMax SGS and gMean gSLOPE sequences.
\item \code{3} uses the BH SLOPE and gMean gSLOPE sequences, also known as SGS Original.
}}

\item{type}{The type of regression to perform. Supported values are: "linear" and "logistic".}

\item{lambda}{The value of \eqn{\lambda}, which defines the level of sparsity in the model. Can be picked using cross-validation (see fit_sgs_cv)}

\item{alpha}{The value of \eqn{\alpha}, defines the convex balance between SLOPE and gSLOPE.}

\item{vFDR}{Defines the desired variable FDR level, which determines the shape of the variable penalties.}

\item{gFDR}{Defines the desired group FDR level, which determines the shape of the group penalties.}

\item{max_iter}{Maximum number of ATOS iterations to perform.}

\item{backtracking}{The backtracking parameter, as defined in Pedregosa et. al. (2018) as \eqn{\tau}.}

\item{max_iter_backtracking}{Maximum number of backtracking line search iterations to perform per global iteration.}

\item{tol}{Convergence threshold for the stopping criteria.}

\item{standardise}{Type of standardisation.
\itemize{
\item \code{l2} standardises the input data to have \eqn{\ell_2} norms of one.
\item \code{l1} standardises the input data to have \eqn{\ell_1} norms of one.
\item \code{sd} standardises the input data to have standard deviation of one.
\item \code{none} no standardisation applied.
}}

\item{intercept}{Logical flag for whether to fit an intercept.}

\item{w_weights}{Optional vector for the group penalty weights. Overrides the penalties from pen_method if specified.}

\item{v_weights}{Optional vector for the variable penalty weights. Overrides the penalties from pen_method if specified.}

\item{x0}{Optional initial vector for \eqn{x_0}.}

\item{u}{Optional initial vector for \eqn{u}.}

\item{verbose}{Logical flag for whether to print fitting information.}
}
\value{
A list containing:
item{beta}{The fitted values from the regression.}
item{x}{The solution to the OPT (See Pedregosa et. al. (2018))}
item{u}{The solution to the dual problem (See Pedregosa et. al. (2018))}
item{z}{The updated values from applying the first proximal operator. (See Pedregosa et. al. (2018))}
item{type}{Indicates which type of regression was performed.}
item{pen_slope}{Vector of the variable penalty sequence.}
item{pen_gslope}{Vector of the group penalty sequence.}
item{lambda}{Value of \eqn{\lambda} used to fit the model.}
item{success}{Logical flag indicating whether ATOS converged, according to item{tol}.}
item{num_it}{Number of iterations performed. If convergence is not reached, this will be item{max_iter}.}
item{certificate}{Final value of convergence criteria.}
item{intercept}{Logical flag indicating whether an intercept was fit.}
}
\description{
Sparse-group SLOPE (SGS) main fitting function.
}
\details{
Fits SGS models using ATOS.
}
\examples{
# specify a grouping structure
groups = c(rep(1:20, each=3),
          rep(21:40, each=4),
          rep(41:60, each=5),
          rep(61:80, each=6),
          rep(81:100, each=7))
# generate data
data = generate_toy_data(p=500, n=400, groups = groups, seed_id=3)
# run SGS 
model = fit_sgs(X = data$X, y = data$y, groups = groups, type="linear", lambda = 1, alpha=0.95, vFDR=0.1, gFDR=0.1, standardise = "l2", intercept = TRUE, verbose=FALSE)
}
\references{
F. Feser, M. Evangelou \emph{Sparse-group SLOPE}, \url{https://github.com/ff1201/sgs}
}
