% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sgs_cv.R
\name{fit_sgs_cv}
\alias{fit_sgs_cv}
\title{fit an SGS model using CV}
\usage{
fit_sgs_cv(
  X,
  y,
  groups,
  pen_method = 1,
  type = "linear",
  nlambda = 20,
  nfolds = 10,
  alpha = 0.95,
  vFDR = 0.1,
  gFDR = 0.1,
  backtracking = 0.7,
  max_iter = 5000,
  max_iter_backtracking = 100,
  tol = 1e-05,
  min_frac = 0.05,
  standardise = "l2",
  intercept = TRUE,
  verbose = FALSE,
  v_weights = NULL,
  w_weights = NULL,
  error_criteria = "mse",
  max_lambda = NULL
)
}
\arguments{
\item{X}{Input matrix of dimensions \eqn{p \times n}.}

\item{y}{Output vector of dimension \eqn{n}.}

\item{groups}{A grouping structure for the input data. Should take the form of a vector of group indices.}

\item{pen_method}{The type of penalty sequences to use.
\itemize{
\item \code{1} uses the vMean SGS and gMean gSLOPE sequences.
\item \code{2} uses the vMax SGS and gMean gSLOPE sequences.
\item \code{3} uses the BH SLOPE and gMean gSLOPE sequences, also known as SGS Original.
}}

\item{type}{The type of regression to perform. Supported values are: "linear" and "logistic".}

\item{nlambda}{The number of pathwise \eqn{\lambda} values to fit.}

\item{nfolds}{The number of folds to use in cross-validation.}

\item{alpha}{The value of \eqn{\alpha}, which defines the convex balance between SLOPE and gSLOPE.}

\item{vFDR}{Defines the desired variable FDR level, which determines the shape of the variable penalties.}

\item{gFDR}{Defines the desired group FDR level, which determines the shape of the group penalties.}

\item{backtracking}{The backtracking parameter, as defined in Pedregosa et. al. (2018) as \eqn{\tau}.}

\item{max_iter}{Maximum number of ATOS iterations to perform.}

\item{max_iter_backtracking}{Maximum number of backtracking line search iterations to perform per global iteration.}

\item{tol}{Convergence threshold for the stopping criteria.}

\item{min_frac}{Defines the termination point of the pathwise solution, so that \eqn{\lambda_\text{min} = min_frac \cdot \lambda_\text{max}}.}

\item{standardise}{Type of standardisation.
\itemize{
\item \code{l2} standardises the input data to have \eqn{\ell_2} norms of one.
\item \code{l1} standardises the input data to have \eqn{\ell_1} norms of one.
\item \code{sd} standardises the input data to have standard deviation of one.
\item \code{none} no standardisation applied.
}}

\item{intercept}{Logical flag for whether to fit an intercept.}

\item{verbose}{Logical flag for whether to print fitting information.}

\item{v_weights}{Optional vector for the variable penalty weights. Overrides the penalties from pen_method if specified.}

\item{w_weights}{Optional vector for the group penalty weights. Overrides the penalties from pen_method if specified.}

\item{error_criteria}{The criteria used to discriminate between models along the path. Supported values are: "mse" (mean squared error) and "mae" (mean absolute error).}

\item{max_lambda}{Optional parameter, which is the value of \eqn{\lambda_\text{max}}, which is used to fit the first model on the path. If not specificed, it is chosen to give the null model.}
}
\value{
A list containing:
item{all_models}{A list of all the models fitted along the path.}
item{fit}{The 1se chosen model, which is a \code{"sgs"} object type.}
item{best_lambda}{The value of \eqn{\lambda} which generated the chosen model.}
item{best_lambda_id}{The path index for the chosen model.}
item{errors}{A table containing fitting information about the models on the path.}
item{type}{Indicates which type of regression was performed.}
}
\description{
Function to fit a pathwise solution of SGS models using k-fold cross-validation.
}
\details{
Fits SGS models under a pathwise solution using ATOS, picking the 1se model as optimum. An example fit is given. Warm starts are implemented.
}
\examples{
# specify a grouping structure
groups = c(rep(1:20, each=3),
          rep(21:40, each=4),
          rep(41:60, each=5),
          rep(61:80, each=6),
          rep(81:100, each=7))
# generate data
data = generate_toy_data(p=500, n=400, groups = groups, seed_id=3)
# run SGS with cross-validation (the proximal functions can be found in utils.R)
cv_model = fit_sgs_cv(X = data$X, y = data$y, groups=groups, type = "linear", nlambda = 20, nfolds=10, alpha = 0.95, vFDR = 0.1, gFDR = 0.1, min_frac = 0.05, standardise="l2",intercept=TRUE,verbose=TRUE)
}
\references{
F. Feser, M. Evangelou \emph{Sparse-group SLOPE}, \url{https://github.com/ff1201/sgs}
}
