% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgsR.R
\name{gain_data}
\alias{gain_data}
\title{Extract the data from Seurat object
Used by ExportsToSGS}
\usage{
gain_data(
  object,
  reductions = NULL,
  markers.file = NULL,
  ident.field = NULL,
  matrix.slot = "scale.data",
  markers.n = 100
)
}
\arguments{
\item{object}{Seurat object}

\item{reductions}{vector of reduction names to export, defaults to all reductions.}

\item{markers.file}{path to file with marker genes.}

\item{ident.field}{cluster.field name of the metadata field containing cell cluster}

\item{matrix.slot}{matrix to use, default is 'scaled data'}

\item{markers.n}{if no markers were supplied, FindAllMarkers is run.
This parameter indicates how many markers to calculate, default is 100}
}
\value{
Return the list of merged df and list of informs
}
\description{
This function used to  gain the data from the seurat object
}
