% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_penalties.R
\name{generate_penalties}
\alias{generate_penalties}
\title{generate penalty sequences for SGS}
\usage{
generate_penalties(gFDR, vFDR, pen_method, groups, alpha)
}
\arguments{
\item{gFDR}{Defines the desired group FDR level, which determines the shape of the group penalties.}

\item{vFDR}{Defines the desired variable FDR level, which determines the shape of the variable penalties.}

\item{pen_method}{The type of penalty sequences to use.
\itemize{
\item \code{1} uses the vMean SGS and gMean gSLOPE sequences.
\item \code{2} uses the vMax SGS and gMean gSLOPE sequences.
\item \code{3} uses the BH SLOPE and gMean gSLOPE sequences, also known as SGS Original.
}}

\item{groups}{A grouping structure for the input data. Should take the form of a vector of group indices.}

\item{alpha}{The value of \eqn{\alpha}, defines the convex balance between SLOPE and gSLOPE.}
}
\value{
A list containing:
item{pen_slope_org}{A vector of the variable penalty sequence.}
item{pen_gslope_org}{A vector of the group penalty sequence.}
}
\description{
Generates variable and group penalties for SGS.
}
\examples{
# specify a grouping structure
groups = c(rep(1:20, each=3),
          rep(21:40, each=4),
          rep(41:60, each=5),
          rep(61:80, each=6),
          rep(81:100, each=7))
# generate sequences
sequences = generate_penalties(gFDR=0.1, vFDR=0.1, pen_method=1, groups=groups, alpha=0.5)

}
\references{
F. Feser, M. Evangelou \emph{Sparse-group SLOPE}, \url{https://github.com/ff1201/sgs}
}
