% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subGroup-methods.R
\name{get.allAncestors}
\alias{get.allAncestors}
\title{Get relevant DAG nodes (ancestors)}
\usage{
get.allAncestors(dag, A)
}
\arguments{
\item{dag}{DAG as permutation matrix}

\item{A}{vector containing the evidence nodes (order: 1 instead of "A")}
}
\value{
relevant DAG
}
\description{
Outputs the relevant subnetwork of the Bayesian network given the evidence
}
