% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subGroup-methods.R
\name{get.allSubGroups}
\alias{get.allSubGroups}
\title{Get All Conditionally Independent Subgroups (CIS)}
\usage{
get.allSubGroups(DAG, evidenceNodes, visualize = F)
}
\arguments{
\item{evidenceNodes}{vector containing the evidence nodes}

\item{BayesNet}{Bayesian network of type bn.fit}
}
\value{
list of all CIS of the Bayesian network given the evidence
}
\description{
Outputs all CISs of the Bayesian network given the evidence
}
