% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subGroup-methods.R
\name{get.subGroup}
\alias{get.subGroup}
\title{Get Conditionally Independent Subgroup (CIS)}
\usage{
get.subGroup(DAG, evidenceNodes, startNode)
}
\arguments{
\item{DAG}{Bayesian network of type bn.fit}

\item{evidenceNodes}{vector containing the evidence nodes}

\item{startNode}{node contained in the subgroup}
}
\value{
all nodes of the corresponding CIS
}
\description{
Outputs the corresponding CIS of a specific node in the Bayesian network given the evidence
}
