% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion_BiDAG_Bestie-methods.R
\name{learn_bn}
\alias{learn_bn}
\title{learn_bn}
\usage{
learn_bn(mydata, bdepar = list(chi = 0.5, edgepf = 2))
}
\arguments{
\item{mydata}{input data for learning}

\item{bdepar}{hyper-parameters for structure learning}
}
\value{
Bayesian network of type SGS
}
\description{
Outputs the Bayesian network DAG and CPTs
}
