% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/noise_estimation.R
\name{noise_estimation}
\alias{noise_estimation}
\title{fits an SGS model using noise estimation}
\usage{
noise_estimation(
  X,
  y,
  groups,
  type = "linear",
  pen_method = 1,
  alpha = 0.95,
  vFDR = 0.1,
  gFDR = 0.1,
  standardise = "l2",
  intercept = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{X}{Input matrix of dimensions \eqn{p \times n}.}

\item{y}{Output vector of dimension \eqn{n}.}

\item{groups}{A grouping structure for the input data. Should take the form of a vector of group indices.}

\item{type}{The type of regression to perform. Supported values are: "linear" and "logistic".}

\item{pen_method}{The type of penalty sequences to use.
\itemize{
\item \code{1} uses the vMean and gMean SGS sequences.
\item \code{2} uses the vMax and gMax SGS sequences.
}}

\item{alpha}{The value of \eqn{\alpha}, defines the convex balance between SLOPE and gSLOPE.}

\item{vFDR}{Defines the desired variable FDR level, which determines the shape of the variable penalties.}

\item{gFDR}{Defines the desired group FDR level, which determines the shape of the group penalties.}

\item{standardise}{Type of standardisation.
\itemize{
\item \code{l2} standardises the input data to have \eqn{\ell_2} norms of one.
\item \code{l1} standardises the input data to have \eqn{\ell_1} norms of one.
\item \code{sd} standardises the input data to have standard deviation of one.
\item \code{none} no standardisation applied.
}}

\item{intercept}{Logical flag for whether to fit an intercept.}

\item{verbose}{Logical flag for whether to print fitting information.}
}
\value{
An object of type \code{"sgs"} containing model fit information.
}
\description{
Fits an SGS model using the noise estimation procedure (Algorithm 2 from SGS paper). This estimates \eqn{\lambda} and then fits the model using the estimated value. It is an alternative approach to cross-validation (\code{\link[=fit_sgs_cv]{fit_sgs_cv()}}). The approach is only compatible with the SGS penalties.
}
\examples{
# specify a grouping structure
groups = c(rep(1:20, each=3),
          rep(21:40, each=4),
          rep(41:60, each=5),
          rep(61:80, each=6),
          rep(81:100, each=7))
# generate data
data = generate_toy_data(p=500, n=400, groups = groups, seed_id=3)
# run noise estimation 
model = noise_estimation(X=data$X, y=data$y, groups=groups)
}
\references{
F. Feser, M. Evangelou \emph{Sparse-group SLOPE}, \url{https://github.com/ff1201/sgs}
}
