% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pbmc_small}
\alias{pbmc_small}
\title{Human PBMC 3k test dataset
A dataset containing the seurat analysis object}
\format{
A seurat object
}
\usage{
pbmc_small
}
\description{
Human PBMC 3k test dataset
A dataset containing the seurat analysis object
}
\keyword{datasets}
