% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.sgs}
\alias{predict.sgs}
\title{predict using a \code{"sgs"} object}
\usage{
\method{predict}{sgs}(fit, predict_X, ...)
}
\arguments{
\item{fit}{Object an object of class \code{"sgs"} from a call to \code{\link[=fit_sgs]{fit_sgs()}}.}

\item{predict_X}{Input data to use for prediction.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
A list containing:
item{response}{The predicted response. In the logistic case, this represents the predicted class probabilities.}
item{class}{The predicted class assignments. Only returned if type = "logistic" in the \code{"sgs"} object.}
}
\description{
Performs prediction from an \code{\link[=fit_sgs]{fit_sgs()}} model fit.
}
\examples{
# specify a grouping structure
groups = c(rep(1:20, each=3),
          rep(21:40, each=4),
          rep(41:60, each=5),
          rep(61:80, each=6),
          rep(81:100, each=7))
# generate data
data = generate_toy_data(p=500, n=400, groups = groups, seed_id=3)
# run SGS 
model = fit_sgs(X = data$X, y = data$y, groups = groups, type="linear", lambda = 1, alpha=0.95, vFDR=0.1, gFDR=0.1, standardise = "l2", intercept = TRUE, verbose=FALSE)
# use predict function
model_predictions = predict(model, predict_X = data$X)
}
\seealso{
\code{\link[=fit_sgs]{fit_sgs()}}

Other SGS-methods: 
\code{\link{plot.sgs_cv}()},
\code{\link{print.sgs}()}
}
\concept{SGS-methods}
