% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subGroup-methods.R
\name{randomBN}
\alias{randomBN}
\title{Create random Bayesian network}
\usage{
randomBN(
  N_nodes,
  N_neighbours = 2,
  nodeDim = 2,
  method = "er",
  uniformCPTs = TRUE,
  visualize = FALSE
)
}
\arguments{
\item{N_nodes}{number of nodes of the network (30 by default)}

\item{N_neighbours}{expected number of neighbours per node (binary by default)}

\item{nodeDim}{number of possible assignments of each node (binary by default)}

\item{method}{Graph type input for randDAG ("er" by default)}

\item{uniformCPTs}{whether CPTs are simulated from a uniform distribution}

\item{visualize}{If true, the network is plotted}
}
\value{
random Bayesian network of type BN()
}
\description{
Creates a random Bayesian networks where the DAG simulation is based on bnlearn
(Generates graphs whose node ordering is given by the order of the labels in the
nodes argument)
}
