% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling-methods.R
\name{sample.normConst}
\alias{sample.normConst}
\title{Importance Sampling (IS) in Bayesian Networks}
\usage{
sample.normConst(
  BayesNet,
  obs,
  s_method = "SGS",
  N_samples = 100,
  relevantSubNet = FALSE,
  plot = TRUE
)
}
\arguments{
\item{BayesNet}{Bayesian network}

\item{obs}{list containing the evidence nodes and associated values}
}
\value{
relevant DAG
}
\description{
Outputs the normalizing constant given some evidence of a Bayesian network
}
