% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atos.R
\name{atos}
\alias{atos}
\title{Adaptive three operator splitting (ATOS)}
\usage{
atos(
  X,
  y,
  loss,
  prox_1,
  prox_2,
  pen_prox_1 = 0.5,
  pen_prox_2 = 0.5,
  max_iter = 5000,
  backtracking = 0.7,
  max_iter_backtracking = 100,
  tol = 1e-05,
  prox_1_opts = NULL,
  prox_2_opts = NULL,
  group_ids,
  standardise = FALSE
)
}
\arguments{
\item{X}{Input matrix of dimensions \eqn{p \times n}.}

\item{y}{Output vector of dimension \eqn{n}.}

\item{loss}{The loss function to use (\eqn{f(x)}). Supported values are: "mse".}

\item{prox_1}{The proximal operator for the first function (\eqn{h(x)}).}

\item{prox_2}{The proximal operator for the second function (\eqn{g(x)}).}

\item{pen_prox_1}{The penalty for the first proximal operator. For the lasso, this would be \eqn{\lambda}. If operator does not include a penalty, set to 1.}

\item{pen_prox_2}{The penalty for the second proximal operator. For the lasso, this would be \eqn{\lambda}. If operator does not include a penalty, set to 1.}

\item{backtracking}{The backtracking parameter, as defined in Pedregosa et. al. (2018) as \eqn{\tau}.}

\item{max_iter_backtracking}{Maximum number of backtracking line search iterations to perform per global iteration.}

\item{tol}{Convergence threshold for the stopping criteria.}

\item{prox_1_opts}{Optional arguments for first proximal operator. For the group lasso, this would be a vector of group IDs.}

\item{prox_2_opts}{Optional arguments for second proximal operator. For the group lasso, this would be a vector of group IDs.#' @param max_iter Maximum number of ATOS iterations to perform.}

\item{standardise}{Logical flag for standardising input data.}
}
\value{
A list containing:\tabular{ll}{
\code{x} \tab The solution to the OPT (See Pedregosa et. al. (2018)) \cr
\tab \cr
\code{u} \tab The solution to the dual problem (See Pedregosa et. al. (2018)) \cr
\tab \cr
\code{success} \tab Logical flag indicating whether ATOS converged, according to \code{tol}. \cr
\tab \cr
\code{num_it} \tab Number of iterations performed. If convergence is not reached, this will be \code{max_iter}. \cr
\tab \cr
}
}
\description{
General function to perform ATOS with penalties other than SGS.
}
\details{
Performs ATOS for other penalties. An example for the sparse-group lasso (SGL) is given.
The algorithm is not symmetrical, but usually the difference between variations are only small numerical values, which are filtered out.
However, both variations should be checked regardless.
}
\examples{
# specify 6 groups of sizes 2, 3, and 4
group <- c(1, 1, 2, 2, 2, 3, 3, 3, 3,
           4, 4, 5, 5, 5, 6, 6, 6, 6)
# set the weight for each group to the square root of the group's size
wt <- rep(c(sqrt(2), sqrt(3), sqrt(4)), 2)
names(wt) <- 1:6
# compute different lambda sequences
lambda.max <- lambdaGroupSLOPE(method="max", fdr=0.1, group=group, wt=wt) 
lambda.mean <- lambdaGroupSLOPE(method="mean", fdr=0.1, group=group, wt=wt) 
lambda.corrected <- lambdaGroupSLOPE(method="corrected", fdr=0.1,
                                     group=group, wt=wt, n.obs=1000)
rbind(lambda.max, lambda.mean, lambda.corrected)
#                      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]
# lambda.max       2.023449 1.844234 1.730818 1.645615 1.576359 1.517427
# lambda.mean      1.880540 1.723559 1.626517 1.554561 1.496603 1.447609
# lambda.corrected 1.880540 1.729811 1.637290 1.568971 1.514028 1.467551

}
\references{
F. Pedregosa, G. Gidel (2018) \emph{Adaptive Three Operator Splitting}, \url{https://proceedings.mlr.press/v80/pedregosa18a.html}
}
