% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subGroup-methods.R
\name{exactInference}
\alias{exactInference}
\title{Exact inference via SupGroupSeparation}
\usage{
exactInference(BayesNet, obs)
}
\arguments{
\item{BayesNet}{Bayesian network of type bn.fit}

\item{obs}{observation}
}
\value{
marginal probability
}
\description{
Outputs the the marginal probability
}
