% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_sgs.R
\name{fit_sgs}
\alias{fit_sgs}
\title{Sparse-group SLOPE (SGS)}
\usage{
fit_sgs(
  X,
  y,
  groups,
  type = "linear",
  lambda,
  alpha,
  FDR,
  gFDR,
  max_iter = 5000,
  backtracking = 0.7,
  max_iter_backtracking = 100,
  tol = 1e-05,
  standardise = FALSE
)
}
\arguments{
\item{X}{Input matrix of dimensions \eqn{p \times n}.}

\item{y}{Output vector of dimension \eqn{n}.}

\item{groups}{A vector for the grouping structure. Each input variable should have a group id.}

\item{type}{The model type. Supported values are: "linear".}

\item{lambda}{Defines the degree of sparsity in the model, as in the lasso.}

\item{alpha}{Controls the level of sparsity between SLOPE and gSLOPE. Acts as in SGL. Must be in the range \eqn{[0,1]}. \code{alpha = 0} returns SLOPE and \code{alpha = 1} returns gSLOPE.}

\item{FDR}{The target false-discovery rate at a variable level. This value is used to define the BH penalty sequence for the variables (\eqn{v}). Must be in the range \eqn{(0,1)}.}

\item{gFDR}{The target false-discovery rate at a group level. This value is used to define the BH penalty sequence for the groups (\eqn{w}). Must be in the range \eqn{(0,1)}.}

\item{max_iter}{Maximum number of ATOS iterations to perform.}

\item{backtracking}{The backtracking parameter, as defined in Pedregosa et. al. (2018) as \eqn{\tau}.}

\item{max_iter_backtracking}{Maximum number of backtracking line search iterations to perform per global iteration.}

\item{tol}{Convergence threshold for the stopping criteria.}

\item{standardise}{Logical flag for standardising input data.}
}
\value{
A list containing:\tabular{ll}{
\code{beta} \tab The estimated coefficients \cr
\tab \cr
\code{selected_var} \tab Indices of non-zero \code{beta} values. \cr
\tab \cr
\code{selected_group} \tab Indices of non-zero groups. \cr
\tab \cr
\code{success} \tab Logical flag indicating whether ATOS converged, according to \code{tol}. \cr
\tab \cr
\code{num_it} \tab Number of iterations performed. If convergence is not reached, this will be \code{max_iter}. \cr
\tab \cr
}
}
\description{
Fit a sparse-group SLOPE model.
}
\details{
Function performs fitting of the sparse-group SLOPE model. The model is fitted using Adaptive Three Operator Splitting (ATOS)
}
\examples{
# specify 6 groups of sizes 2, 3, and 4
group <- c(1, 1, 2, 2, 2, 3, 3, 3, 3,
           4, 4, 5, 5, 5, 6, 6, 6, 6)
# set the weight for each group to the square root of the group's size
wt <- rep(c(sqrt(2), sqrt(3), sqrt(4)), 2)
names(wt) <- 1:6
# compute different lambda sequences
lambda.max <- lambdaGroupSLOPE(method="max", fdr=0.1, group=group, wt=wt) 
lambda.mean <- lambdaGroupSLOPE(method="mean", fdr=0.1, group=group, wt=wt) 
lambda.corrected <- lambdaGroupSLOPE(method="corrected", fdr=0.1,
                                     group=group, wt=wt, n.obs=1000)
rbind(lambda.max, lambda.mean, lambda.corrected)
#                      [,1]     [,2]     [,3]     [,4]     [,5]     [,6]
# lambda.max       2.023449 1.844234 1.730818 1.645615 1.576359 1.517427
# lambda.mean      1.880540 1.723559 1.626517 1.554561 1.496603 1.447609
# lambda.corrected 1.880540 1.729811 1.637290 1.568971 1.514028 1.467551

}
\references{
F. Pedregosa, G. Gidel (2018) \emph{Adaptive Three Operator Splitting}, \url{https://proceedings.mlr.press/v80/pedregosa18a.html}
}
