% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark-methods.R
\name{makeAllPlots}
\alias{makeAllPlots}
\title{Visualize benchmark results}
\usage{
makeAllPlots(
  resultTable1,
  resultTable2,
  resultTable3,
  resultTable4,
  labelBP = c("1", "2", "3", "4"),
  labelHead = c("1", "2", "3", "4"),
  fileName = "Test",
  width = 7,
  height = 4.1,
  niceBoxPlot = FALSE,
  retP = FALSE
)
}
\arguments{
\item{resultTable1}{result table 1 of function benchmarkMultipleNets()}

\item{resultTable2}{result table 2 of function benchmarkMultipleNets()}

\item{resultTable3}{result table 3 of function benchmarkMultipleNets()}

\item{resultTable4}{result table 4 of function benchmarkMultipleNets()}

\item{labelBP}{label}

\item{fileName}{file name}

\item{width}{figure width}

\item{height}{figure height}

\item{niceBoxPlot}{if TRUE, only a fraction of points (150) are plotted in boxplot}

\item{retP}{if TRUE, a plot is returned for summary plot, otherwise no ruturn}
}
\value{
plot of benchmark results
}
\description{
Outputs plots of benchmark results for different inference methods
}
