% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subGroupSampling-methods.R
\name{sample.subGroupSampling}
\alias{sample.subGroupSampling}
\title{Sub Group Sampling (SGS) in Bayesian Networks}
\usage{
sample.subGroupSampling(
  BayesNet,
  obs,
  N_samples = 100,
  relevantSubNet = FALSE,
  plot = TRUE
)
}
\arguments{
\item{BayesNet}{Bayesian network}

\item{obs}{list containing the evidence nodes and associated values}
}
\value{
relevant DAG
}
\description{
Outputs the normalizing constant given some evidence of a Bayesian network
}
