% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs-package.R
\docType{package}
\name{sgs-package}
\alias{sgs-package}
\alias{_PACKAGE}
\title{sgs: Sparse-Group SLOPE: Adaptive Bi-Level Selection with FDR Control}
\description{
Implementation of Sparse-group SLOPE: Adaptive bi-level with FDR-control (Feser et al. (2023) \href{https://arxiv.org/abs/2305.09467}{arXiv:2305.09467}). Linear and logistic regression models are supported, both of which can be fit using k-fold cross-validation. Dense and sparse input matrices are supported. In addition, a general adaptive three operator splitting (ATOS) implementation is provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ff1201/sgs}
  \item Report bugs at \url{https://github.com/ff1201/sgs/issues}
}

}
\author{
\strong{Maintainer}: Fabio Feser \email{ff120@ic.ac.uk} (\href{https://orcid.org/0009-0007-3088-9727}{ORCID})

Authors:
\itemize{
  \item Marina Evangelou (\href{https://orcid.org/0000-0003-0789-8944}{ORCID})
}

}
\keyword{internal}
