% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs-package.R
\docType{package}
\name{sgs-package}
\alias{sgs}
\alias{sgs-package}
\title{sgs: Sparse-Group SLOPE: Adaptive Bi-Level Selection with FDR Control}
\description{
Implementation of Sparse-group SLOPE (SGS) (Feser and Evangelou (2023) \doi{10.48550/arXiv.2305.09467}) models. Linear and logistic regression models are supported, both of which can be fit using k-fold cross-validation. Dense and sparse input matrices are supported. In addition, a general Adaptive Three Operator Splitting (ATOS) (Pedregosa and Gidel (2018) \doi{10.48550/arXiv.1804.02339}) implementation is provided. Group SLOPE (gSLOPE) (Brzyski et al. (2019) \doi{10.1080/01621459.2017.1411269}) and group-based OSCAR models (Feser and Evangelou (2024) \doi{10.48550/arXiv.2405.15357}) are also implemented. All models are available with strong screening rules (Feser and Evangelou (2024) \doi{10.48550/arXiv.2405.15357}) for computational speed-up.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ff1201/sgs}
  \item Report bugs at \url{https://github.com/ff1201/sgs/issues}
}

}
\author{
\strong{Maintainer}: Fabio Feser \email{ff120@ic.ac.uk} (\href{https://orcid.org/0009-0007-3088-9727}{ORCID})

}
\keyword{internal}
