% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sgs-package.R
\docType{package}
\name{sgs-package}
\alias{sgs-package}
\alias{_PACKAGE}
\title{sgs: Sparse-group SLOPE: adaptive bi-level selection with FDR control}
\description{
Implementation of Sparse-group SLOPE: Adaptive bi-level with FDR-control and applications in genetics (Bogdan et al. (2015) \doi{10/gfgwzt}). Linear and logistic regression models are supported, both of which can be fit using k-fold cross-validation. Dense and sparse input matrices are supported. In addition, a general adaptive three operator splitting (ATOS) implementation is provided.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ff1201/sgs}
  \item Report bugs at \url{https://github.com/ff1201/sgs/issues}
}

}
\author{
\strong{Maintainer}: Fabio Feser \email{ff120@ic.ac.uk} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

Authors:
\itemize{
  \item Marina Evangelou
}

}
\keyword{internal}
